+++ ruoyi-system/src/main/java/com/ruoyi/system/domain/work/WorkFullyMechanizedStatsJu.java
@@ -7,6 +7,7 @@ import lombok.EqualsAndHashCode;
 import org.apache.commons.lang3.builder.ToStringBuilder;
 import org.apache.commons.lang3.builder.ToStringStyle;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 /**
@@ -44,19 +45,19 @@ public class WorkFullyMechanizedStatsJu extends BaseEntity {
 
     /** 一班产量 */
     @Excel(name = "一班产量")
-    private Integer outputShift1;
+    private BigDecimal outputShift1;
 
     /** 二班产量 */
     @Excel(name = "二班产量")
-    private Integer outputShift2;
+    private BigDecimal outputShift2;
 
     /** 三班产量 */
     @Excel(name = "三班产量")
-    private Integer outputShift3;
+    private BigDecimal outputShift3;
 
     /** 原班产量 */
     @Excel(name = "原班产量")
-    private Integer originalOutput;
+    private BigDecimal originalOutput;
 
     /** 累计产量 */
     @Excel(name = "累计产量")
